tidy up stationService
